Move tests into jest
