+++ .travis.yml
@@ -7,5 +7,5 @@ cache: yarn
 script:
   - (cd packages/catalyst && yarn tsc --noEmit)
   - (cd packages/catalyst-client && yarn tsc --noEmit)
-  - (cd packages/catalyst && ./tests.js)
+  - (cd packages/catalyst && yarn test)
   - (cd packages/catalyst-client && yarn test)
