+++ packages/catalyst/package.json
@@ -27,6 +27,9 @@
   "bugs": {
     "url": "https://github.com/friendsoftheweb/catalyst/issues"
   },
+  "jest": {
+    "globalSetup": "<rootDir>/test-setup.js"
+  },
   "dependencies": {
     "@babel/core": "^7.4.4",
     "@babel/plugin-proposal-class-properties": "^7.4.4",
