+++ packages/catalyst/test-project/__tests__/symbol-polyfill.test.js
@@ -1,0 +1,14 @@
+const assertFileContains = require('../assertFileContains');
+
+test('build process polyfills symbols', async () => {
+  await assertFileContains({
+    file: 'symbol-polyfill.js',
+    content:
+      '__webpack_require__(/*! core-js/modules/es.symbol */ "../node_modules/core-js/modules/es.symbol.js");'
+  });
+
+  await assertFileContains({
+    file: 'symbol-polyfill.js',
+    content: "var testSymbol = Symbol('testSymbol');"
+  });
+});
